First Commit
Updated pom.xml
* Update bukkit to 1.13.2
* Changed mcstats to bstats
