--- config.yml
@@ -1,90 +1,0 @@
-config-version: 1
-properties:
-  # Show debug messages.
-  debug-messages: false
-
-  # Help the developer know how popular the plugin is.
-  use-metrics: true
-  
-  # Check a chunk upon load (ChunkLoadEvent).
-  check-chunk-load: false
-  
-  # Check a chunk upon unload (ChunkUnloadEvent).
-  check-chunk-unload: false
-  
-  # Check a chunk when a mob spawns (CreatureSpawnEvent).
-  watch-creature-spawns: true
-  
-  # Radius of surrounding chunks to check.
-  check-surrounding-chunks: 1
-  
-  # When a chunk is loaded, recheck it periodically. 
-  active-inspections: true
-  
-  # How often, in seconds, to check the chunk.
-  inspection-frequency: 300
-  
-  # Notify players in that chunk if stuff has been culled. 
-  notify-players: false
-
-# Spawn reasons to cull on.
-spawn-reasons:
-  NATURAL: true
-  JOCKEY: true
-  CHUNK_GEN: true
-  SPAWNER: true
-  EGG: true
-  SPAWNER_EGG: true
-  LIGHTNING: true
-  # These are deprecated, might not work in the future. 
-  BED: true
-  BUILD_SNOWMAN: true
-  BUILD_IRONGOLEM: true
-  BUILD_WITHER: true
-  VILLAGE_DEFENSE: true
-  VILLAGE_INVASION: true
-  BREEDING: true
-  SLIME_SPLIT: true
-  REINFORCEMENTS: true
-  CUSTOM: true
-  DEFAULT: true
-
-entities:
-  ANIMAL: 50
-  MONSTER: 50
-  NPC: 50
-  OTHER: 500
-#  WATER_MOB: 5
-#  AMBIENT: 5
-#  CREEPER: 10
-#  SHEEP: 10
-#  OCELOT: 10
-#  GIANT: 10
-#  SLIME: 10
-#  GHAST: 10
-#  PIG_ZOMBIE: 10
-#  ENDERMAN: 10
-#  SILVERFISH: 10
-#  MAGMA_CUBE: 10
-#  CHICKEN: 10
-#  SQUID: 10
-#  WOLF: 10
-#  MUSHROOM_COW: 10
-#  SNOWMAN: 10
-#  VILLAGER: 10
-#  SPIDER: 10
-#  SKELETON: 10
-#  ZOMBIE: 10
-#  BLAZE: 10
-#  CAVE_SPIDER: 10
-#  IRON_GOLEM: 10
-#  WITHER: 10
-#  PIG: 10
-#  HORSE: 10
-#  WITCH: 10
-
-# Exclude these worlds from limits.  
-excluded-worlds: []
-
-messages:
-  removedEntites: "&7Removed %s %s in your chunk."
